added getText methods to meele and added the broad blade back in

+++ dungeon-game-core/src/com/dungeon/game/item/weapon/part/Part.java
@@ -96,7 +96,7 @@ public abstract class Part extends Item implements Cloneable{
 		return new ArrayList<Effect>();
 	};
 	
-	public abstract void setStats(int level);
+	public abstract void setStats(float level);
 	
 	
 	//does fuck all. VERY IMPORTANT. (actually though: if you delete it the game will be kill)
